# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: ECDAT API
  version: "1.0.0"
  description: >
    Enterprise Cryptographic Discovery & Analysis Tool (SIH26164) API contract.
    Phase 0: this contract is implemented by an in-memory stub (see
    backend/api/), not a real detection engine. Examples on Finding/Scan/etc
    are hand-built, internally consistent with the Mosca risk formula, and
    explicitly NOT measured detector output (no engine exists yet).
servers:
  - url: /api/v1
    description: Relative to backend host; full generated paths (used by contract_diff) are under /api/v1.
paths:
...
  /api/v1/estate/trend:
    get:
      operationId: getEstateTrend
      summary: Historical estate risk trend over time
      tags: [estate]
      parameters:
        - name: days
          in: query
          required: false
          schema: { type: integer, default: 30 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EstateTrend"
...
  /api/v1/residue:
    get:
      operationId: listResidueClusters
      summary: List crypto debt residue clusters across scans and targets
      tags: [debt]
      parameters:
        - name: state
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/ResidueClusterState"
        - name: targetId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResidueCluster"
  /api/v1/residue/{id}:
    get:
      operationId: getResidueCluster
      summary: Get details of a specific crypto debt residue cluster
      tags: [debt]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResidueCluster"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: patchResidueCluster
      summary: Transition state of a residue cluster (open -> promoted | excluded | accepted)
      tags: [debt]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResidueClusterPatch"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResidueCluster"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
...
  responses:
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorDetail"
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorDetail"
  schemas:
    ErrorDetail:
      type: object
      required: [error, message]
      properties:
        error: { type: string }
        message: { type: string }
...
    EstateTrendPoint:
      type: object
      required: [date, avgRiskScore, criticalCount, totalFindings]
      properties:
        date: { type: string }
        avgRiskScore: { type: number }
        criticalCount: { type: integer }
        totalFindings: { type: integer }
    EstateTrend:
      type: object
      required: [days, points]
      properties:
        days: { type: integer }
        points:
          type: array
          items: { $ref: "#/components/schemas/EstateTrendPoint" }
...
    ResidueOccurrence:
      type: object
      required: [artifactHash, path, range]
      properties:
        artifactHash: { type: string }
        path: { type: string }
        range:
          type: array
          items: { type: integer }
    ResidueClusterState:
      type: string
      enum: [open, promoted, excluded, accepted]
    ResidueCluster:
      type: object
      required: [id, contentHash, signalTypes, magnitude, occurrences, state, firstSeen, lastSeen]
      properties:
        id: { type: string }
        contentHash: { type: string }
        signalTypes:
          type: array
          items: { type: string }
        magnitude: { type: number }
        occurrences:
          type: array
          items: { $ref: "#/components/schemas/ResidueOccurrence" }
        state: { $ref: "#/components/schemas/ResidueClusterState" }
        justification: { type: [string, "null"] }
        owner: { type: [string, "null"] }
        firstSeen: { type: string, format: date-time }
        lastSeen: { type: string, format: date-time }
    ResidueClusterPatch:
      type: object
      required: [state]
      properties:
        state: { $ref: "#/components/schemas/ResidueClusterState" }
        justification: { type: [string, "null"] }
        owner: { type: [string, "null"] }
...